refactor component folder structure
