+++ my-app/package.json
@@ -20,7 +20,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "ts-check": "tsc --project ./tsconfig.json"
   },
   "eslintConfig": {
     "extends": [
