+++ my-app/src/components/onedayWeather/index.ts
@@ -1,0 +1,1 @@
+export * from "./OnedayWeather"
