+++ my-app/src/components/todayWeather/index.ts
@@ -1,0 +1,1 @@
+export * from "./TodayWeather"
