+++ my-app/src/components/weeklyWeather/index.ts
@@ -1,0 +1,1 @@
+export * from "./WeeklyWeather"
